backend/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
mypy==0.940
mypy-extensions==0.4.3
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing_extensions==4.1.1
